requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
argh==0.31.3
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.1.0
backoff==1.11.1
bcrypt==4.3.0
blinker==1.9.0
Bottleneck==1.4.2
Brotli==1.0.9
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.14.6
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
durationpy==0.9
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.10
fastapi-cli==0.0.7
filelock==3.17.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2==2.1.5
google-auth==2.38.0
googleapis-common-protos==1.68.0
greenlet==1.1.2
grpcio==1.62.2
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface_hub==0.29.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
Jinja2==3.1.5
jiter==0.6.1
jsonpatch==1.33
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==32.0.1
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-text-splitters==0.3.6
langsmith==0.2.11
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==5.2.0
mypy-extensions==1.0.0
narwhals==1.27.1
numexpr==2.10.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.19.2
openai==1.65.1
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.14
overrides==7.7.0
packaging==24.2
pandas==2.2.3
Pillow==8.3.2
pip==25.0
pkgutil_resolve_name==1.3.10
posthog==3.6.5
propcache==0.2.1
protobuf==4.25.3
pulsar-client==3.5.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydantic-settings==2.8.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pypdf==5.3.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pytz-deprecation-shim==0.1.0.post0
pyu2f==0.1.5
PyYAML==6.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.11.3
rpds-py==0.22.3
rsa==4.9
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.46.0
streamlit==1.42.2
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tornado==6.1
tqdm==4.67.1
typer==0.15.2
typer-slim==0.15.2
typing_extensions==4.12.2
typing-inspect==0.9.0
typing_utils==0.1.0
tzdata==2025.1
tzlocal==4.0.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchdog==2.1.6
watchfiles==1.0.4
websocket-client==1.8.0
websockets==15.0
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
### missing requirements
langchain-openai
sentence-transformers 
